requirements.txt
pandas
google-genai
tqdm
scrapetube
langdetect
